# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gold Trading AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signal:
    get:
      operationId: getSignal
      tags: [scalper]
      summary: Get current AI scalping signal
      parameters:
        - name: timeframe
          in: query
          required: false
          schema:
            type: string
            enum: [1m, 5m]
            default: 5m
      responses:
        "200":
          description: Current scalping signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScalperSignalResponse"
...
  /telegram/configure:
    post:
      operationId: configureTelegram
      tags: [scalper]
      summary: Configure Telegram bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramConfig"
      responses:
        "200":
          description: Configuration saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramConfigResponse"
...
  /intraday/signal:
    get:
      operationId: getIntradaySignal
      tags: [intraday]
      summary: Get current AI intraday signal
      parameters:
        - name: timeframe
          in: query
          required: false
          schema:
            type: string
            enum: [15m, 30m, 1h]
            default: 15m
      responses:
        "200":
          description: Current intraday signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntradaySignalResponse"
...
  /smc/signal:
    get:
      operationId: getSmcSignal
      tags: [smc]
      summary: Get Smart Money Concepts AI signal
      parameters:
        - name: timeframe
          in: query
          required: false
          schema:
            type: string
            enum: [15m, 1h]
            default: 15m
      responses:
        "200":
          description: SMC signal with full analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmcSignalResponse"
...
components:
  schemas:
...
    ScalperIndicators:
      type: object
      properties:
        rsi:
          type: number
        ema20:
          type: number
        ema50:
          type: number
        macdLine:
          type: number
        macdSignal:
          type: number
        macdHistogram:
          type: number
        atr:
          type: number
      required: [rsi, ema20, ema50, macdLine, macdSignal, macdHistogram, atr]
    ScalperSignalResponse:
      type: object
      properties:
        signal:
          type: string
          enum: [BUY, SELL, HOLD]
        confidence:
          type: number
        entry:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        trend:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        marketMode:
          type: string
          enum: [TRENDING, SIDEWAYS]
        signalLabel:
          type: string
          description: Human-readable label for the signal type (e.g. TREND FOLLOWING SELL, REVERSAL BUY)
        timeframe:
          type: string
        indicators:
          $ref: "#/components/schemas/ScalperIndicators"
        timestamp:
          type: string
          format: date-time
      required: [signal, confidence, entry, stopLoss, takeProfit, trend, marketMode, timeframe, indicators, timestamp]
...
    IntradayIndicators:
      type: object
      properties:
        rsi:
          type: number
        ema20:
          type: number
        ema50:
          type: number
        macdLine:
          type: number
        macdSignal:
          type: number
        macdHistogram:
          type: number
        atr:
          type: number
        support:
          type: number
        resistance:
          type: number
      required: [rsi, ema20, ema50, macdLine, macdSignal, macdHistogram, atr, support, resistance]
    IntradaySignalResponse:
      type: object
      properties:
        signal:
          type: string
          enum: [BUY, SELL, HOLD]
        confidence:
          type: number
        entry:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        trend:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        timeframe:
          type: string
        holdingPeriod:
          type: string
        indicators:
          $ref: "#/components/schemas/IntradayIndicators"
        timestamp:
          type: string
          format: date-time
      required: [signal, confidence, entry, stopLoss, takeProfit, trend, timeframe, holdingPeriod, indicators, timestamp]
...
    TelegramConfig:
      type: object
      properties:
        botToken:
          type: string
        chatId:
          type: string
      required: [botToken, chatId]
    TelegramConfigResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    SmcSetup:
      type: object
      properties:
        bos:
          type: boolean
        choch:
          type: boolean
        liquidityGrab:
          type: boolean
        orderBlock:
          type: boolean
        fvg:
          type: boolean
      required: [bos, choch, liquidityGrab, orderBlock, fvg]
    SmcConfidenceBreakdown:
      type: object
      properties:
        trendAlignment:
          type: number
        liquidityConfirmation:
          type: number
        orderBlockReaction:
          type: number
        structureConfirmation:
          type: number
        indicatorSupport:
          type: number
        total:
          type: number
      required: [trendAlignment, liquidityConfirmation, orderBlockReaction, structureConfirmation, indicatorSupport, total]
    SmcSignalResponse:
      type: object
      properties:
        signal:
          type: string
          enum: [BUY, SELL, HOLD]
        confidence:
          type: number
        entry:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        marketStructure:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        htfTrend:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL]
        setup:
          $ref: "#/components/schemas/SmcSetup"
        confidenceBreakdown:
          $ref: "#/components/schemas/SmcConfidenceBreakdown"
        setupDescription:
          type: string
        holdingPeriod:
          type: string
        rsi:
          type: number
        timeframe:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [signal, confidence, entry, stopLoss, takeProfit, marketStructure, htfTrend, setup, confidenceBreakdown, setupDescription, holdingPeriod, rsi, timeframe, timestamp]
...