# lib/api-zod/src/generated/types/smcConfidenceBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Trading AI API
 * OpenAPI spec version: 0.1.0
 */

export interface SmcConfidenceBreakdown {
  trendAlignment: number;
  liquidityConfirmation: number;
  orderBlockReaction: number;
  structureConfirmation: number;
  indicatorSupport: number;
  total: number;
}