# lib/api-zod/src/generated/types/smcSetup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gold Trading AI API
 * OpenAPI spec version: 0.1.0
 */

export interface SmcSetup {
  bos: boolean;
  choch: boolean;
  liquidityGrab: boolean;
  orderBlock: boolean;
  fvg: boolean;
}